src/main/java/de/fnortheim/spring5recipes/controllers/IndexController.java
package de.fnortheim.spring5recipes.controllers;

import de.fnortheim.spring5recipes.services.RecipeService;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.GetMapping;

/**
 * created by sebastian on Apr, 2019
 */
@Slf4j
@Controller
public class IndexController {

    private RecipeService recipeService;

    public IndexController(RecipeService recipeService) {
        this.recipeService = recipeService;
    }

    @GetMapping({"", "/", "/index"})
    public String getIndexPage(Model model) {
        log.debug("getting index page");

        model.addAttribute("recipes", recipeService.getRecipes());

        return "index";
    }
}
